refactor(gamesection): separate gamesectionheader component / change to server component
